% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ginteff.R
\name{as.data.frame.ginteff}
\alias{as.data.frame.ginteff}
\title{Convert a ginteff object to a tidy data frame}
\usage{
\method{as.data.frame}{ginteff}(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{a \code{ginteff} object}

\item{row.names}{ignored}

\item{optional}{ignored}

\item{...}{ignored}
}
\description{
Convert a ginteff object to a tidy data frame
}
