% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ginteff.R
\name{vcov.ginteff}
\alias{vcov.ginteff}
\title{Variance-covariance matrix of the AIE vector}
\usage{
\method{vcov}{ginteff}(object, ...)
}
\arguments{
\item{object}{a \code{ginteff} object}

\item{...}{ignored}
}
\description{
Returns the delta-method variance-covariance matrix of the average
interaction effects stored in \code{object$aie}. Useful for chaining with
downstream tools (e.g. \pkg{multcomp}, \pkg{car}::deltaMethod).
}
